Made a shit ton of prefabs so that they can be used for the actual level with colliders attached, added EmergencyRigidbody(TM) for player and enemies so that they at least collide, player dies when hit by toaster. BLADE TRIGGER DOESN'T REACT EVEN WHEN CLEARLY HITTING SOMETHING!

+++ Assets/Code/Player/PlayerCombat.cs
@@ -11,7 +11,7 @@ public class PlayerCombat : MonoBehaviour {
     public GameObject meleeHitColliderObject;
 
     public float meleeHitRadius = 1.0f;
-    public float knifeDamageAmount = 100.0f;
+    public int knifeDamageAmount = 100;
 
     // Controlled thru UIControl.
     public bool canAttack = true;
